Fixed an oversight with the face shadow softness
I used a regular if else statement to test the softness to be above a certain threshold. This controlled whether or not the shader should step() the shadow region or if it should use it as it comes from the shadow_ndotl() function.
This wouldn't normally be a problem except that this is something that should ONLY be used when the material is toggled as a face material.